fixes #18132 - hooks in --noop delete pulp.conf (#468)
this workaround is only needed for smart proxies, main katello
overwrites the pulp.conf. Noop check added to both hooks.

+++ hooks/post/29-create_package_httpd_conf.rb
@@ -8,12 +8,14 @@
 # This hook creates placeholders files so the package does not put them in
 # place anymore.
 #
-%w(pulp.conf).each do |file|
-  if !File.file?(File.join("/etc/httpd/conf.d/", file))
-    File.open(File.join('/etc/httpd/conf.d', file), 'w') do |f|
-      f.write("# This file is managed by the foreman-installer, do not alter.")
+if !app_value(:noop)
+  %w(pulp.conf).each do |file|
+    if !File.file?(File.join("/etc/httpd/conf.d/", file))
+      File.open(File.join('/etc/httpd/conf.d', file), 'w') do |f|
+        f.write("# This file is managed by the foreman-installer, do not alter.")
+      end
+    else
+      logger.info "#{file} is already present, skipping"
     end
-  else
-    logger.info "#{file} is already present, skipping"
   end
 end
